move constraint tests to separate file.  fix confict/depend error message tests
